added support for double for EVERYTHING

+++ tests/pass/MultAssign.java
@@ -1,7 +1,7 @@
 package pass;
 
 public class MultAssign {
-    public int multAssign(int x, int y) {
+    public double multAssign(double x, double y) {
         y *= x;
         return y;
     }
